+++ SocketDefault.h
@@ -46,12 +46,11 @@ namespace Neko
             // see ISocket for comments
             
             SocketDefault() = delete;
-            
             SocketDefault(const Net::INetSocket& socket);
             
-            virtual long GetPacketBlocking(void* buffer, const uint32 length, const uint32& timeout) const override;
+            virtual long GetPacketBlocking(void* buffer, const ulong length, const uint32& timeout) const override;
             
-            virtual long SendAllPacketsWait(const void* buffer, const uint32 length, const uint32& timeout) const override;
+            virtual long SendAllPacketsWait(const void* buffer, const ulong length, const uint32& timeout) const override;
             
             virtual void Close() override;
